+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [WP-CLI] Improved documentation.
 
 ### Fixed
+- Typos in admin screens.
 - [WP-CLI] Typos in documentation.
 
 ## [2.0.0] - 2020-09-29